% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initPortableStreams.R
\name{getState}
\alias{getState}
\title{Return one or all state vectors, either from original position or
current position.}
\usage{
getState(stream, origin = FALSE)
}
\arguments{
\item{stream}{Optional. Selects a particular stream's state.  If
no value is supplied, then the list including all of the
stream states is supplied.}

\item{origin}{If FALSE (default), return the current, updated
state of one or all random stream. If TRUE, return the initial
stream states.}
}
\value{
Integer index value of currently selected stream
}
\description{
The environment variable currentStates is a list of states of
random generators. In most use cases, users will not need to use
this function because the details are handled by useStream.  Users
may need to explicitly access the stream states if they want
to extract that information and then pass it along to other
functions that do not use the built-in R random generator framework.
}
\examples{
mySeeds <- seedCreator(500, 5, file = "mySeeds.rds", seed = 123123)
setSeeds(mySeeds, run = 17)
runif(2)
getCurrentStream()
getState(origin=TRUE)
getState(stream = 2)
getState()
useStream(2)
runif(2)
}
\author{
Paul E. Johnson <pauljohn@ku.edu>
}
