% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seedCreator.R
\name{update.portableSeeds}
\alias{update.portableSeeds}
\title{Add initializing streams for more runs to a portableSeeds object}
\usage{
\method{update}{portableSeeds}(object, more, file = NULL, ...)
}
\arguments{
\item{object}{A portableSeeds object created by seedCreator}

\item{more}{An integer indicating the number of additional runs
for which seeds are to be created.}

\item{file}{A quoted character string representing file name into
which the portableSeeds object is to be placed.}

\item{...}{Additional arguments, none of which are put to any use}
}
\value{
A \code{portableSeeds} object that includes the original set, with
\code{more} seeds added to the end. The entirety of this returned set
should equal the set that would have been generated by seedCreator if
the same number of seeds had been requested.
}
\description{
After seedCreator has created a collection of seeds, sometimes it
is necessary to conduct additional runs that require more
initializing values. This function fulfills that need.
}
\examples{
require(parallel)
projSeeds <- seedCreator(2000, 3, seed = 123456, file = "someSeeds.rds")
projSeeds2 <- update(projSeeds, more = 19, file = "somePlusSeeds.rds")
identical(projSeeds[[1]], projSeeds2[[1]])
identical(projSeeds[[2000]], projSeeds2[[2000]])
projSeeds2[[2001]]
projSeedsLast <- projSeeds[[2000]][[3]]
projSeedsNext <- parallel::nextRNGStream(projSeedsLast)
identical(projSeedsNext, projSeeds2[[2001]][[1]])
}
\author{
Paul E. Johnson <pauljohn@ku.edu>
}
