% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initPortableStreams.R
\name{initPortableStreams}
\alias{initPortableStreams}
\title{The name previously used for the function
\code{\link{setSeeds}}.}
\usage{
initPortableStreams(projSeeds, run, verbose = FALSE)
}
\arguments{
\item{projSeeds}{Required. Either an object of class portableSeeds
(created by \code{\link{seedCreator}}) or a text string giving the
name of an R saved file of the appropriate format (created by
the seedCreator function).}

\item{run}{Integer indicating which element from the portable seed
collection is to be selected}

\item{verbose}{Optional. Print out the state of the current
generator. Default = FALSE.}
}
\value{
nothing is returned. This function is used for the side
    effect of setting three objects in the global environment, the
    startStates (list), currentStates (list), and currentStream
    (an integer).
}
\description{
This is replaced by \code{\link{setSeeds}}, which does exactly the
same thing, but with a name that is more immediatly understandable
to users.  This is now deprecated, will be removed.
}
\seealso{
\code{\link{setSeeds}} which is the replacement for this
    function.
}
\author{
Paul E. Johnson <pauljohn@ku.edu>
}
