% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/initPortableStreams.R
\name{initPortableStreams}
\alias{initPortableStreams}
\title{The name previously used for the function
\code{\link{setSeeds}}.}
\usage{
initPortableStreams(projSeeds, run, verbose = FALSE)
}
\description{
This is replaced by setSeeds, which does exactly the same thing,
but with a name that is more immediatly understandable to users.
This is now deprecated, will be removed.
}

